Fix qt5 build and fix windows build

+++ libshviotqt/tests/serialportsocket/mockserialport.cpp
@@ -50,8 +50,9 @@ qint64 MockSerialPort::writeData(const char *data, qint64 len)
 
 qint64 MockSerialPort::readData(char *data, qint64 maxlen)
 {
-	auto len = std::min(m_dataToRead.size(), maxlen);
-	for(qint64 i = 0; i < len; ++i) {
+	// FIXME: get rid of the casts here when dropping Qt5 support
+	auto len = std::min(static_cast<int>(m_dataToRead.size()), static_cast<int>(maxlen));
+	for(int i = 0; i < len; ++i) {
 		data[i] = m_dataToRead[i];
 	}
 	m_dataToRead = m_dataToRead.mid(len);
